renamed files to fit with our coding guidelines
