+++ QAChat/QA_Bot/api_interface.py
@@ -2,8 +2,8 @@
 # SPDX-FileCopyrightText: 2023 Emanuel Erben
 # SPDX-FileCopyrightText: 2023 Felix Nützel
 
-from QAChat.QA_Bot.QABot import QABot
-import QAChat.Slack_Bot.BaseAgent as BaseAgent
+from QAChat.QA_Bot.qa_bot import QABot
+import QAChat.Slack_Bot.base_agent as BaseAgent
 
 
 class APIInterface:
